% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettystatsBy.R
\name{statsBy}
\alias{statsBy}
\title{Extending the statsBy Funtionality of the Psych Package.}
\usage{
statsBy(data = NULL, group = NULL, alpha = 0.05, var_names = NULL, ...)
}
\arguments{
\item{data}{Dataframe only including Variables to be included in correlation table.}

\item{group}{Clustering variable, such as a personID}

\item{alpha}{significance Level that is passed to the psych::statsBy function.
This significance level is used for all outputs provided by the original
psych::statsBy function, but is not used for the functionality provided
by this specific package.}

\item{var_names}{a list of names that are used for pretty printing}

\item{...}{Refer to the psych::statsBy documentation for additional parameters.}
}
\value{
The function returns a psych::statsBy object that retains all
   functionality of the original object, while extending the functionality.
   The function automatically calculates confidence intervals with the alpha
   levels of 0.05, 0.01, and 0.001 and provides formatted correlation tables
   with indications of significance. All original calls and exact confidence
   intervals for the alpha level provided to the function can still be
   obtained via the regular functionality of the psych::statsBy function.
}
\description{
Extending the statsBy Funtionality of the Psych Package.
}
\examples{
# create the object
statsByObject <- statsBy(df[,vars], df$userID, var_names=vars)

# extract information
print(statsByObject) # To explore the original functionality
print(statsByObject$pretty.within) # Full within- cluster correlation matrix
print(statsByObject$pretty.between) # Full between- cluster correlation matrix
print(statsByObject$pretty.combined) # top half are within-correlations, bottom half are between correlations.


}
