% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettystatsBy.R
\name{statsBy}
\alias{statsBy}
\title{Extending the statsBy Funtionality of the Psych Package}
\usage{
statsBy(
  data = NULL,
  group = NULL,
  alpha = 0.05,
  var_names = NULL,
  pretty_alphas = c(0.05, 0.01, 0.001),
  ...
)
}
\arguments{
\item{data}{Dataframe only including Variables to be included in correlation table.}

\item{group}{Clustering variable, such as a personID}

\item{alpha}{significance Level that is passed to the psych::statsBy function.
This significance level is used for all outputs provided by the original
psych::statsBy function, but is not used for the functionality provided
by this specific package.}

\item{var_names}{a list of names that are used for pretty printing}

\item{pretty_alphas}{you may provide a list of 3 alpha levels that are used
to calculate the confidence interval for the prettystatsBy output.?p}

\item{...}{Refer to the psych::statsBy documentation for additional parameters.}
}
\value{
The function returns a psych::statsBy object that retains all
   functionality of the original object, while extending the functionality.
   The function automatically calculates confidence intervals with the
   alpha leves specified in pretty_alphas and provides formatted correlation
   tables with indications of significance. All original calls and exact
   confidence intervals for the single alpha level provided to the function can
   still be obtained via the regular functionality of the psych::statsBy function.

use `Object$pretty`: provides a list of three dataframes ready for exporting.

use `Object$pretty.ciw` or `Object$pretty.cib` to obtain lists of
confidence intercals for the three provided alpha levels.
}
\description{
Extending the statsBy Funtionality of the Psych Package
}
\examples{
# create the object (example 1)
statsByObject <- statsBy(df[,vars], df$userID, var_names=vars)
# create the object (example 2)
statsByObject <- statsBy(df, 'userID', var_names=colnames(df))

# extract information
print(statsByObject) # To explore the original functionality
print(statsByObject$pretty # Access all correlation matrices this package adds.
print(statsByObject$pretty$within) # Full within- cluster correlation matrix
print(statsByObject$pretty$between) # Full between- cluster correlation matrix
print(statsByObject$pretty$combined) # top half are within-correlations, bottom half are between correlations.
}
